Extract the block fetcher's peer selection logic for data columns so it can be used in both by range and by root requests
