+++ beacon-chain/p2p/interfaces.go
@@ -117,4 +117,5 @@ type DataColumnsHandler interface {
 	CustodyGroupCountFromPeer(peer.ID) uint64
 	AdmissibleCustodyGroupsPeers([]peer.ID) ([]peer.ID, error)
 	AdmissibleCustodySamplingPeers([]peer.ID) ([]peer.ID, error)
+	AdmissiblePeersForCustodyGroups([]peer.ID, map[uint64]bool) (map[peer.ID]map[uint64]bool, map[uint64][]peer.ID, []string, error)
 }
